# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LearnerProfile {
  name: string;
  role: string;
  department: string;
  streakDays: number;
}

export interface LearningStats {
  learningHours: number;
  completedCourses: number;
  inProgressCourses: number;
  readinessScore: number;
}

export interface LearningCourse {
  id: number;
  title: string;
  category: string;
  description: string;
  instructor: string;
  durationHours: number;
  level: string;
  rating: number;
  userProgress: number;
  accent: string;
}

export interface Competency {
  id: number;
  name: string;
  category: string;
  currentLevel: number;
  targetLevel: number;
  impact: string;
}

export interface RoadmapStep {
  id: number;
  courseId: number;
  title: string;
  category: string;
  estimatedHours: number;
  status: string;
  position: number;
}

export interface LearningActivity {
  id: number;
  title: string;
  detail: string;
  kind: string;
  occurredAt: string;
}

export interface LearningDashboard {
  learner: LearnerProfile;
  stats: LearningStats;
  competencies: Competency[];
  recommendedCourses: LearningCourse[];
  activity: LearningActivity[];
  weeklyHours: number[];
}

export type LearningSettingsCoachStyle = typeof LearningSettingsCoachStyle[keyof typeof LearningSettingsCoachStyle];


export const LearningSettingsCoachStyle = {
  supportive: 'supportive',
  concise: 'concise',
  challenging: 'challenging',
} as const;

export interface LearningSettings {
  /**
     * @minimum 1
     * @maximum 20
     */
  weeklyStudyGoalHours: number;
  coachStyle: LearningSettingsCoachStyle;
}

export type LearningSettingsUpdateCoachStyle = typeof LearningSettingsUpdateCoachStyle[keyof typeof LearningSettingsUpdateCoachStyle];


export const LearningSettingsUpdateCoachStyle = {
  supportive: 'supportive',
  concise: 'concise',
  challenging: 'challenging',
} as const;

export interface LearningSettingsUpdate {
  /**
     * @minimum 1
     * @maximum 20
     */
  weeklyStudyGoalHours: number;
  coachStyle: LearningSettingsUpdateCoachStyle;
}

export interface ProgressUpdate {
  /**
     * @minimum 0
     * @maximum 100
     */
  percentComplete: number;
}

export interface CourseProgress {
  courseId: number;
  percentComplete: number;
}

export interface LearningCoachInput {
  /**
     * @minLength 2
     * @maxLength 2000
     */
  question: string;
  /** @maxLength 4000 */
  context?: string;
}

export interface LearningCoachReply {
  answer: string;
}