# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learning/settings:
    get:
      operationId: getLearningSettings
      tags: [learning]
      summary: Get learner learning preferences
      responses:
        "200":
          description: Saved study goal and coach style
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSettings"
    patch:
      operationId: updateLearningSettings
      tags: [learning]
      summary: Update learner learning preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningSettingsUpdate"
      responses:
        "200":
          description: Updated study goal and coach style
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSettings"
        "400":
          description: Invalid learning preferences
  /learning/courses/{courseId}/progress:
    patch:
      operationId: updateCourseProgress
      tags: [learning]
      summary: Update progress for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated course progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
        "400":
          description: Invalid progress value
        "404":
          description: Course not found
  /gemini/learning-coach:
    post:
      operationId: askLearningCoach
      tags: [ai]
      summary: Ask the AI learning coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningCoachInput"
      responses:
        "200":
          description: Coaching response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningCoachReply"
        "400":
          description: Invalid request
        "429":
          description: Too many coaching requests
        "502":
          description: AI provider returned an error
components:
  schemas:
...
    LearningSettings:
      type: object
      required: [weeklyStudyGoalHours, coachStyle]
      properties:
        weeklyStudyGoalHours: { type: integer, minimum: 1, maximum: 20 }
        coachStyle:
          type: string
          enum: [supportive, concise, challenging]
    LearningSettingsUpdate:
      type: object
      required: [weeklyStudyGoalHours, coachStyle]
      properties:
        weeklyStudyGoalHours: { type: integer, minimum: 1, maximum: 20 }
        coachStyle:
          type: string
          enum: [supportive, concise, challenging]
    ProgressUpdate:
      type: object
      required: [percentComplete]
      properties:
        percentComplete: { type: integer, minimum: 0, maximum: 100 }
    CourseProgress:
      type: object
      required: [courseId, percentComplete]
      properties:
        courseId: { type: integer }
        percentComplete: { type: integer }
    LearningCoachInput:
      type: object
      required: [question]
      properties:
        question: { type: string, minLength: 2, maxLength: 2000 }
        context: { type: string, maxLength: 4000 }
    LearningCoachReply:
      type: object
      required: [answer]
      properties:
        answer: { type: string }