# lib/api-zod/src/generated/types/learningSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningSettingsCoachStyle } from './learningSettingsCoachStyle';

export interface LearningSettings {
  /**
     * @minimum 1
     * @maximum 20
     */
  weeklyStudyGoalHours: number;
  coachStyle: LearningSettingsCoachStyle;
}