# lib/api-zod/src/generated/types/learningSettingsUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningSettingsUpdateCoachStyle } from './learningSettingsUpdateCoachStyle';

export interface LearningSettingsUpdate {
  /**
     * @minimum 1
     * @maximum 20
     */
  weeklyStudyGoalHours: number;
  coachStyle: LearningSettingsUpdateCoachStyle;
}